[omicron-package] sanitize cargo subprocess env vars (#7945)
