+++ dev-tools/common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "dev-tools-common"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+# Please read the note in `dev-tools/xtask/Cargo.toml` before adding a new
+# dependency!
