+++ package/Cargo.toml
@@ -13,6 +13,7 @@ anyhow.workspace = true
 camino.workspace = true
 cargo_metadata.workspace = true
 clap.workspace = true
+dev-tools-common.workspace = true
 futures.workspace = true
 hex.workspace = true
 illumos-utils.workspace = true
